src/admin_panel/serializers.py
from rest_framework import serializers
from .models import Requests
from django.contrib.auth.models import User


class RequestSerializer(serializers.ModelSerializer):
    class Meta:
        model = Requests
        fields = ['id', 'name', 'phone', 'created_at']


class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = ['id', 'username', 'email']
