requirements.txt
asgiref==3.9.1
Django==5.2.6
django-cors-headers==4.8.0
django-modeltranslation==0.19.17
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
pillow==11.3.0
PyJWT==2.10.1
sqlparse==0.5.3
